LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/nested-query-containers-expected.txt


PASS test1 - inline - 0b000
PASS test2 - inline - 0b001
PASS test3 - inline - 0b010
PASS test4 - inline - 0b011
PASS test5 - inline - 0b100
PASS test6 - inline - 0b101
PASS test7 - inline - 0b110
PASS test8 - inline - 0b111
PASS test9 - contents - 0b000
PASS test10 - contents - 0b001
PASS test11 - contents - 0b010
PASS test12 - contents - 0b011
PASS test13 - contents - 0b100
PASS test14 - contents - 0b101
PASS test15 - contents - 0b110
PASS test16 - contents - 0b111
FAIL test17 - table-cell - 0b000 assert_equals: clientHeight expected 2 but got 5
FAIL test18 - table-cell - 0b001 assert_equals: clientHeight expected 2 but got 5
FAIL test19 - table-cell - 0b010 assert_equals: clientHeight expected 2 but got 5
FAIL test20 - table-cell - 0b011 assert_equals: clientHeight expected 2 but got 5
FAIL test21 - table-cell - 0b100 assert_equals: clientHeight expected 2 but got 5
FAIL test22 - table-cell - 0b101 assert_equals: clientHeight expected 2 but got 5
FAIL test23 - table-cell - 0b110 assert_equals: clientHeight expected 2 but got 5
FAIL test24 - table-cell - 0b111 assert_equals: clientHeight expected 2 but got 5
FAIL test25 - table - 0b000 assert_equals: clientHeight expected 4 but got 8
FAIL test26 - table - 0b001 assert_equals: clientHeight expected 4 but got 8
FAIL test27 - table - 0b010 assert_equals: clientHeight expected 4 but got 8
FAIL test28 - table - 0b011 assert_equals: clientHeight expected 4 but got 8
FAIL test29 - table - 0b100 assert_equals: clientHeight expected 4 but got 8
FAIL test30 - table - 0b101 assert_equals: clientHeight expected 4 but got 8
FAIL test31 - table - 0b110 assert_equals: clientHeight expected 4 but got 8
FAIL test32 - table - 0b111 assert_equals: clientHeight expected 4 but got 8
